Move some image related methods & struct to smaller files
